Update filterevents.rst
I was having issues testing the filter, which works fine running under my account or root - but was failing for the "www-data" user (permission issue). I'd propose that the '-u www-data' be added so users can test the script under the user ZM would call it. Also, the default FILTER_EXECUTE_INTERVAL was 60 on my fresh 1.30 install.